requirements.txt
Flask==3.1.0
python-dotenv==1.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
Flask-Migrate==4.1.0